requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
Django==3.0.6
django-cors-headers==3.2.1
django-jsonfield==1.4.0
django-mysql==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2020.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
